% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutTree.R
\name{cutTree}
\alias{cutTree}
\title{cutTree}
\usage{
cutTree(tree.obj, cut.point)
}
\arguments{
\item{tree.obj}{Tree object, on which the cut will be performed with the given cut point.}

\item{cut.point}{(integer) non-negative numeric; Node ID of parent node at where you want to cut the tree (has to be the parent node of two leafs)}
}
\description{
Cut a single split from your tree (only possible on the lowest level of the tree).
}
\details{
Please note that it is only possible to cut at the lowest level. If you want to cut on a higher level, you have to cut multiple times.
}
\examples{
cutTree(tree.obj = tree, cut.point = 3)
}
