% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{doPrediction}
\alias{doPrediction}
\title{doPrediction}
\usage{
doPrediction(tree.obj, predict.data, pred.class)
}
\arguments{
\item{tree.obj}{a tree object of class InitTreeR}

\item{predict.data}{a data.frame object with test data which should get predicted.}

\item{pred.class}{an boolean for showing the predicted class in the output (optional: default = FALSE)}
}
\description{
Predict the target attribute for the new data of a test set.
}
\examples{
doPrediction(tree.obj = tree, dataset = dat)
}
