% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initTree.R
\name{initTree}
\alias{initTree}
\title{initTree}
\usage{
initTree(data, v.target, auto, theta)
}
\arguments{
\item{v.target}{column name of the target variable in your dataset.}

\item{auto}{default FALSE. If TRUE the tree builds up automatically.}

\item{theta}{non-negative float or double; Improvement restriction when the tree is built automatically (auto = TRUE). Stops if the current best improvement is less than theta.}

\item{dataset}{a data.frame object for the following model building.}
}
\value{
A tree object of class InitTreeR
}
\description{
The function builds a tree object by generating the root node.
}
\details{
This function is a method to charge the best partitions as well as their related improvements.
}
\examples{
tree <- initTree(dataset = data, v.target =”income”, auto = FALSE)
}
