% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPlot.R
\name{plotTree}
\alias{plotTree}
\title{plotTree}
\usage{
plotTree(tree.obj)
}
\arguments{
\item{tree.obj}{A tree object of class InitTreeR}
}
\description{
Plots a tree object, returned from splittree() for example, on the current graphics device.
}
\details{
Plot the tree (realised with Partykit (Toolkit for Recursive Partytioning)).
}
\examples{
plotTree(tree.obj = tree)

}
