% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPlot.R
\name{printTree}
\alias{printTree}
\title{printTree}
\usage{
printTree(tree.obj)
}
\arguments{
\item{tree.obj}{A tree object of class InitTreeR}
}
\description{
Print a tree object, returned from splittree() for example, (as a text tree).
}
\examples{
printTree (tree.obj = tree)
}
