% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printPlot.R
\name{splitOptions}
\alias{splitOptions}
\title{splitOptions}
\usage{
splitOptions(tree.obj, split.point)
}
\arguments{
\item{tree.obj}{A tree object of class InitTreeR}

\item{split.point}{(integer) non-negative numeric; Node ID of terminal node where you want to perform your split.}
}
\description{
Returns a list of improvements for the (next) split at the given split point. The list is sorted by improvements in descending order.
}
\examples{
splitOptions(tree.obj = tree, split.point = 1)
}
