% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitTree.R
\name{splitTree}
\alias{splitTree}
\title{splitTree}
\usage{
splitTree(tree.obj, split.point, split.feature, max.surrogates, custom.split)
}
\arguments{
\item{tree.obj}{The tree object, on which the split will be performed with the given split point.}

\item{split.point}{(integer) non-negative numeric; Node ID of terminal node where you want to perform your split.}

\item{split.featurecharacter}{or string; Column name of the attribute which should be used for the split.}

\item{max.surrogates(Optional)}{(Integer) non-negative; the maximum number of surrogates, which will be calculated for the split. (Default is 3)}

\item{custom.split(Optional)}{(Integer) custom point to split (only for numeric attributes)}
}
\description{
Performs a new binary split at a selected splitpoint.
}
\details{
Missing values are handled via Surrogate Splits. In most details, it follows „Leo Breiman et. al. (1984). “quite closely.
}
\examples{
tree <- splitTree(tree.obj = tree, split.point = 1, split.feature = “income”)
}
